+++ ChangeLog.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.2.7
+* Fix bug in swap `check_address` fn
+
+## 0.2.6
+* Change address size (64 => 66) in swap params
+* Fix BE dpath endianness in swap params
+
+## 0.2.5
+* Fix displaying of UI warning in swap mode 
+
+## 0.2.4
+* Add swap support
+* Bump SDK version
+
 ## 0.2.1
 
 * Update SDK